requirements.txt
py-cord[voice]>=2.4.0
python-dotenv>=1.0.0